Fix build under NetBSD
Explicitly cast to int the parameter passed to isdigit to fix a warning
(and thus a build error in developer mode) under NetBSD-current/i386.

Monotone-Parent: 3b4a1ad38ceec38f9f50a521bdb427ddd811f1ab
Monotone-Revision: 57390d33d0ae757defbf29ceb4a26f62ac30cc31

Monotone-Author: jmmv@NetBSD.org
Monotone-Date: 2008-03-16T15:40:01

+++ atf-c/tp_main.c
@@ -229,7 +229,7 @@ parse_rflag(const char *arg, int *value)
 {
     atf_error_t err;
 
-    if (strlen(arg) != 1 || !isdigit(arg[0])) {
+    if (strlen(arg) != 1 || !isdigit((int)arg[0])) {
         err = usage_error("Invalid value for -r; must be a single digit.");
         goto out;
     }
